requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
prettytable==3.7.0
protobuf==4.23.2
wcwidth==0.2.6
